backend/requirements.txt
asgiref==3.11.1
attrs==26.1.0
cffi==2.0.0
cryptography==46.0.7
diff-match-patch==20241021
Django==6.0
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-debug-toolbar==6.3.0
django-extensions==4.1
django-filter==25.2
django-import-export==4.4.0
django-js-asset==3.1.2
django-redis==6.0.0
django-simple-history==3.11.0
django-simpleui==2026.1.13
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
pillow==12.2.0
pycparser==3.0
PyJWT==2.12.1
PyMySQL==1.1.2
PyYAML==6.0.3
redis==7.4.0
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
tablib==3.9.0
uritemplate==4.2.0
